--- launch/myahrs_driver.launch
@@ -1,7 +1,0 @@
-<launch>
-  <node pkg="myahrs_driver" type="myahrs_driver" name="myahrs_driver">
-    <param name="port" value="/dev/ttyACM0" />
-    <param name="baud_rate" value="115200" />
-  </node>
-  <node pkg="rviz" type="rviz" name="rviz" args="-d $(find myahrs_driver)/rviz_cfg/imu_test.rviz" />
-</launch>
